backend/requirements.txt
Django==3.2.19
djangorestframework==3.12.4
djoser==2.1.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
django-filter==2.4.0
drf-extra-fields==3.0.3
Pillow==9.5.0
python-dotenv==0.19.2
django-cors-headers==3.13.0